chore: split rpc.spec into smaller chunks
